% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extract.R
\name{add_scan_numbers}
\alias{add_scan_numbers}
\title{Add a \code{scan} (acquisition number) column to a chromatogram tibble by matching
retention time per file. \code{meta} must carry id + path columns.}
\usage{
add_scan_numbers(df, meta)
}
\description{
Add a \code{scan} (acquisition number) column to a chromatogram tibble by matching
retention time per file. \code{meta} must carry id + path columns.
}
