% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_filters.R
\name{apply_filters_spectra}
\alias{apply_filters_spectra}
\title{Apply the global filter \code{f} to a Spectra object.}
\usage{
apply_filters_spectra(sp, f)
}
\description{
Apply the global filter \code{f} to a Spectra object.
}
