% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extract.R
\name{bin_peaks}
\alias{bin_peaks}
\title{Bin long peak data into an rt x m/z grid for heatmap / surface display.}
\usage{
bin_peaks(df, rt_bin = 10, mz_bin = 1, aggfun = max)
}
\description{
Bin long peak data into an rt x m/z grid for heatmap / surface display.
}
