% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_palettes.R
\name{brewer_colorscale}
\alias{brewer_colorscale}
\title{A plotly colorscale (list of \link{position, color}) from a sequential palette.}
\usage{
brewer_colorscale(palette = "YlOrRd", n = 9, invert = FALSE)
}
\description{
A plotly colorscale (list of \link{position, color}) from a sequential palette.
}
