% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_palettes.R
\name{brewer_named}
\alias{brewer_named}
\title{Named color vector mapping the levels of a grouping variable to brewer colors.}
\usage{
brewer_named(levels, palette = "Set1")
}
\description{
Named color vector mapping the levels of a grouping variable to brewer colors.
}
