% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_palettes.R
\name{brewer_seq}
\alias{brewer_seq}
\title{A sequential ColorBrewer ramp function for continuous fills (heatmaps/3D).}
\usage{
brewer_seq(palette = "YlOrRd", invert = FALSE)
}
\arguments{
\item{palette}{a sequential ColorBrewer name OR a viridisLite option
(viridis/magma/plasma/inferno/cividis/mako/rocket/turbo)}

\item{invert}{reverse the scale (so it runs light -> dark)}
}
\description{
A sequential ColorBrewer ramp function for continuous fills (heatmaps/3D).
}
