% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extract.R
\name{build_msexp}
\alias{build_msexp}
\title{Build an MsExperiment from included files, injecting our sample metadata.}
\usage{
build_msexp(files_df)
}
\description{
Build an MsExperiment from included files, injecting our sample metadata.
}
