% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extract.R
\name{chrom_to_df}
\alias{chrom_to_df}
\title{Convert an (M/X)Chromatograms object to a tidy tibble for ggplot.}
\usage{
chrom_to_df(chr, meta, labels = NULL)
}
\description{
Convert an (M/X)Chromatograms object to a tidy tibble for ggplot.
}
