% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_filters.R
\name{combined_ranges}
\alias{combined_ranges}
\title{Combined data ranges across the included files (for input hints).}
\usage{
combined_ranges(files_df)
}
\description{
Combined data ranges across the included files (for input hints).
}
