% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extract.R
\name{extract_precursors}
\alias{extract_precursors}
\title{Precursor ions (rt, precursor m/z, scan) for MS>1 spectra in a file (DDA map).}
\usage{
extract_precursors(path)
}
\description{
Precursor ions (rt, precursor m/z, scan) for MS>1 spectra in a file (DDA map).
}
