% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reactive.R
\name{PLOTLY_EVENTS}
\alias{PLOTLY_EVENTS}
\title{The plotly events every interactive plot registers (and that the modules
query via \code{event_data}).}
\usage{
PLOTLY_EVENTS
}
\description{
The plotly events every interactive plot registers (and that the modules
query via \code{event_data}).
}
