% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extract.R
\name{polarity_label}
\alias{polarity_label}
\title{Polarity label from the integer code Spectra uses (0 neg, 1 pos, -1 unknown).}
\usage{
polarity_label(x)
}
\description{
Polarity label from the integer code Spectra uses (0 neg, 1 pos, -1 unknown).
}
