% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_export.R
\name{save_gg}
\alias{save_gg}
\title{Save a ggplot to file using the user's export settings.}
\usage{
save_gg(gg, file, settings)
}
\arguments{
\item{gg}{a ggplot object}

\item{file}{destination path}

\item{settings}{rv$settings list (format/width/height/units/dpi)}
}
\description{
Save a ggplot to file using the user's export settings.
}
