% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemons.R
\name{set_daemons}
\alias{set_daemons}
\title{(Re)size the mirai daemon pool used by the async file readers.}
\usage{
set_daemons(n = .default_daemons())
}
\arguments{
\item{n}{number of daemons (>= 1).}
}
\value{
the daemon count actually set.
}
\description{
(Re)size the mirai daemon pool used by the async file readers.
}
