% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsVisGUI-package.R
\docType{package}
\name{xcmsVisGUI-package}
\alias{xcmsVisGUI}
\alias{xcmsVisGUI-package}
\title{xcmsVisGUI: interactive raw LC-MS viewer (Shiny)}
\description{
A local desktop Shiny app for visualising raw LC-MS data (TIC/BPC, EICs,
spectra, 2D/3D MS maps, DDA precursors). Launch with \code{\link[=run_app]{run_app()}}.
}
\details{
Most cross-package calls are \code{::}-qualified (see the \verb{fct_*} files); only the
functions used bare in the UI/glue code are imported here. The
RforMassSpectrometry stack (Spectra/MsExperiment/xcms/...) is used via \code{::}.
}
\author{
\strong{Maintainer}: Jan Stanstrup \email{stanstrup@gmail.com}

Authors:
\itemize{
  \item Jan Stanstrup \email{stanstrup@gmail.com}
}

}
\keyword{internal}
